% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno_Frankenstein}
\alias{anno_Frankenstein}
\title{Annotated Frankenstein Data Frame
This data frame provides the annotated text of Frankenstein. Using token shows each token and part of speech with other variables. Using enity shows the specific entities such as countries or names. Document shows each document that the text is split into.}
\format{
\subsection{\code{anno_Frankenstein}}{

3 elements
\describe{
\item{token}{A tibble with 85810 rows and 10 columns}
\item{entity}{A tibble with 1643 rows anf 6 columns}
\item{document}{A tibble with 6387 rows and 1 column}
...
}
}
}
\source{
annotate using cleanNLP on Frankestein text
}
\usage{
anno_Frankenstein
}
\description{
Annotated Frankenstein Data Frame
This data frame provides the annotated text of Frankenstein. Using token shows each token and part of speech with other variables. Using enity shows the specific entities such as countries or names. Document shows each document that the text is split into.
}
\keyword{datasets}
