% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frankenstein.R
\docType{data}
\name{Creature}
\alias{Creature}
\title{A Shiny Creature}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Creature
}
\value{
An R6 object
}
\description{
A Shiny Creature
}
\section{Methods}{

\describe{
  \item{\code{as_fresh}}{run the app as fresh}
  \item{\code{dispose}}{remove old saved states}
  \item{\code{list_states}}{list all saved states}
  \item{\code{options}}{shiny options}
  \item{\code{revive}}{revive an old app from last states or specific one}
  \item{\code{server}}{server object}
  \item{\code{ui}}{ui object}
}
}

\keyword{datasets}
