% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Frankenstein}
\alias{Frankenstein}
\title{My data}
\format{
\subsection{\code{Frankenstein}}{

A data frame with 7357 rows and 4 columns:
\describe{
\item{line}{line number}
\item{text}{the text}
\item{section}{the section number}
\item{section_type}{the type of section: letter or chapter}
...
}
}
}
\source{
\url{https://www.gutenberg.org/files/84/84-h/84-h.htm}
}
\usage{
Frankenstein
}
\description{
This is more descriptive text about my data.
Even more info is provided here ...
}
\keyword{datasets}
