% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surgery.R
\name{monitor}
\alias{monitor}
\title{Monitor a Shiny App}
\usage{
monitor(session, intervalMs, dispose = TRUE, folder = "shiny_bookmarks")
}
\arguments{
\item{session}{the session object}

\item{intervalMs}{interval in millisecond between each monitoring}

\item{dispose}{wether or not to dispose of the non-last states}

\item{folder}{the folder where the states are kept}
}
\description{
Monitor a Shiny App
}
